Jasmine Plott Turtle Exercises
This is the actual file that I would like merged and posted on the course page and to count as my actual submission for my turtules.  For the first one, I messed up by not giving the file the correct name and by not putting the title in quotes, which I think will cause problems later on.  I'm sorry about the inconvenience!